Move Undo in the hamburger menu.
No need to force users who do not like
this option to see all the time in the
headerbar an insensitive button. Add a
menu entry instead; ideally, in a menu
dedicated to game actions, but for now
in the hamburger menu. And only if the
option is set to true, of course. That
makes the window layout better anyway.