Remove Preferences.
The Preferences dialog is quite
small, and it does not increase
fun at all. Now with #7 solved,
proposing a change of the speed
is mostly useless (there's only
accessibility that may be taken
in consideration, but then that
would not be a great solution);
the Undo functionality does not
feel right in a game randomized
at each move; and the option to
show congrats is conceptually a
big mistake. So just remove it.

Closes #8.